+++ optimade/models/links.py
@@ -64,7 +64,7 @@ MUST be one of these values: 'child', 'root', 'external', 'providers'.""",
     )
 
     aggregate: Optional[Aggregate] = StrictField(
-        "ok",
+        Aggregate.OK,
         description="""A string indicating whether a client that is following links to aggregate results from different OPTIMADE implementations should follow this link or not.
 This flag SHOULD NOT be indicated for links where `link_type` is not `child`.