+++ optimade/models/structures.py
@@ -67,9 +67,9 @@ class Species(BaseModel):
     - **Examples**:
         - `[ {"name": "Ti", "chemical_symbols": ["Ti"], "concentration": [1.0]} ]`: any site with this species is occupied by a Ti atom.
         - `[ {"name": "Ti", "chemical_symbols": ["Ti", "vacancy"], "concentration": [0.9, 0.1]} ]`: any site with this species is occupied by a Ti atom with 90 % probability, and has a vacancy with 10 % probability.
-        - `[ {"name": "BaCa", "chemical_symbols": ["vacancy", "Ba", "Ca"], "concentration": [0.05, 0.45, 0.5], "mass": 88.5} ]`: any site with this species is occupied by a Ba atom with 45 % probability, a Ca atom with 50 % probability, and by a vacancy with 5 % probability. The mass of this site is (on average) 88.5 a.m.u.
-        - `[ {"name": "C12", "chemical_symbols": ["C"], "concentration": [1.0], "mass": 12.0} ]`: any site with this species is occupied by a carbon isotope with mass 12.
-        - `[ {"name": "C13", "chemical_symbols": ["C"], "concentration": [1.0], "mass": 13.0} ]`: any site with this species is occupied by a carbon isotope with mass 13.
+        - `[ {"name": "BaCa", "chemical_symbols": ["vacancy", "Ba", "Ca"], "concentration": [0.05, 0.45, 0.5], "mass": [0.0, 137.327, 40.078]} ]`: any site with this species is occupied by a Ba atom with 45 % probability, a Ca atom with 50 % probability, and by a vacancy with 5 % probability. The mass of this site is (on average) 88.5 a.m.u.
+        - `[ {"name": "C12", "chemical_symbols": ["C"], "concentration": [1.0], "mass": [12.0]} ]`: any site with this species is occupied by a carbon isotope with mass 12.
+        - `[ {"name": "C13", "chemical_symbols": ["C"], "concentration": [1.0], "mass": [13.0]} ]`: any site with this species is occupied by a carbon isotope with mass 13.
         - `[ {"name": "CH3", "chemical_symbols": ["C"], "concentration": [1.0], "attached": ["H"], "nattached": [3]} ]`: any site with this species is occupied by a methyl group, -CH3, which is represented without specifying precise positions of the hydrogen atoms.
 
     """
@@ -106,9 +106,10 @@ Note that concentrations are uncorrelated between different site (even of the sa
         queryable=SupportLevel.OPTIONAL,
     )
 
-    mass: Optional[float] = OptimadeField(
+    mass: Optional[List[float]] = OptimadeField(
         None,
-        description="""If present MUST be a float expressed in a.m.u.""",
+        description="""If present MUST be a list of floats expressed in a.m.u.
+Elements denoting vacancies MUST have masses equal to 0.""",
         unit="a.m.u.",
         support=SupportLevel.OPTIONAL,
         queryable=SupportLevel.OPTIONAL,
@@ -143,8 +144,10 @@ Note: With regards to "source database", we refer to the immediate source being 
             raise ValueError(f"{v} MUST be in {EXTENDED_CHEMICAL_SYMBOLS}")
         return v
 
-    @validator("concentration")
-    def validate_concentration(cls, v, values):
+    @validator("concentration", "mass")
+    def validate_concentration_and_mass(cls, v, values, field):
+        if not v:
+            return v
         if values.get("chemical_symbols"):
             if len(v) != len(values["chemical_symbols"]):
                 raise ValueError(
@@ -154,7 +157,7 @@ Note: With regards to "source database", we refer to the immediate source being 
             return v
 
         raise ValueError(
-            "Could not validate concentration as 'chemical_symbols' is missing/invalid."
+            "Could not validate {field.name} as 'chemical_symbols' is missing/invalid."
         )
 
     @validator("attached", "nattached")
@@ -562,7 +565,7 @@ Species can represent pure chemical elements, virtual-crystal atoms representing
     - `concentration`: list of float (REQUIRED)
     - `attached`: list of strings (REQUIRED)
     - `nattached`: list of integers (OPTIONAL)
-    - `mass`: float (OPTIONAL)
+    - `mass`: list of floats (OPTIONAL)
     - `original_name`: string (OPTIONAL).
 
 - **Requirements/Conventions**:
@@ -595,7 +598,8 @@ Species can represent pure chemical elements, virtual-crystal atoms representing
           The implementation MUST include either both or none of the `attached` and `nattached` keys, and if they are provided, they MUST be of the same length.
           Furthermore, if they are provided, the `structure_features` property MUST include the string `site_attachments`.
 
-        - **mass**: OPTIONAL. If present MUST be a float expressed in a.m.u.
+        - **mass**: OPTIONAL. If present MUST be a list of floats, with the same length as `chemical_symbols`, providing element masses expressed in a.m.u.
+          Elements denoting vacancies MUST have masses equal to 0.
 
         - **original_name**: OPTIONAL. Can be any valid Unicode string, and SHOULD contain (if specified) the name of the species that is used internally in the source database.
 
@@ -611,9 +615,9 @@ Species can represent pure chemical elements, virtual-crystal atoms representing
 - **Examples**:
     - `[ {"name": "Ti", "chemical_symbols": ["Ti"], "concentration": [1.0]} ]`: any site with this species is occupied by a Ti atom.
     - `[ {"name": "Ti", "chemical_symbols": ["Ti", "vacancy"], "concentration": [0.9, 0.1]} ]`: any site with this species is occupied by a Ti atom with 90 % probability, and has a vacancy with 10 % probability.
-    - `[ {"name": "BaCa", "chemical_symbols": ["vacancy", "Ba", "Ca"], "concentration": [0.05, 0.45, 0.5], "mass": 88.5} ]`: any site with this species is occupied by a Ba atom with 45 % probability, a Ca atom with 50 % probability, and by a vacancy with 5 % probability. The mass of this site is (on average) 88.5 a.m.u.
-    - `[ {"name": "C12", "chemical_symbols": ["C"], "concentration": [1.0], "mass": 12.0} ]`: any site with this species is occupied by a carbon isotope with mass 12.
-    - `[ {"name": "C13", "chemical_symbols": ["C"], "concentration": [1.0], "mass": 13.0} ]`: any site with this species is occupied by a carbon isotope with mass 13.
+    - `[ {"name": "BaCa", "chemical_symbols": ["vacancy", "Ba", "Ca"], "concentration": [0.05, 0.45, 0.5], "mass": [0.0, 137.327, 40.078]} ]`: any site with this species is occupied by a Ba atom with 45 % probability, a Ca atom with 50 % probability, and by a vacancy with 5 % probability. The mass of this site is (on average) 88.5 a.m.u.
+    - `[ {"name": "C12", "chemical_symbols": ["C"], "concentration": [1.0], "mass": [12.0]} ]`: any site with this species is occupied by a carbon isotope with mass 12.
+    - `[ {"name": "C13", "chemical_symbols": ["C"], "concentration": [1.0], "mass": [13.0]} ]`: any site with this species is occupied by a carbon isotope with mass 13.
     - `[ {"name": "CH3", "chemical_symbols": ["C"], "concentration": [1.0], "attached": ["H"], "nattached": [3]} ]`: any site with this species is occupied by a methyl group, -CH3, which is represented without specifying precise positions of the hydrogen atoms.""",
         support=SupportLevel.SHOULD,
         queryable=SupportLevel.OPTIONAL,